Begin plot script clean up
