+++ arg_scripts/plot_args.py
@@ -1,0 +1,34 @@
+empty_props = {
+    'plot_dict': None,
+    'time': None,
+    'sim_num': None,
+    'erlang_dict': None,
+    'num_requests': None,
+    'num_cores': None,
+    'erlang': None,
+    'network': None,
+    'date': None,
+
+    'color_list': ['#024de3', '#00b300', 'orange', '#6804cc', '#e30220'],
+    'style_list': ['solid', 'dashed', 'dotted', 'dashdot'],
+    'marker_list': ['o', '^', 's', 'x'],
+    'x_tick_list': [10, 100, 200, 300, 400],
+}
+
+empty_plot_dict = {
+    'erlang_list': [],
+    'blocking_list': [],
+    'lengths_list': [],
+    'hops_list': [],
+    'net_util_list': [],
+    'active_reqs_list': [],
+    'block_req_list': [],
+    'req_num_list': [],
+    'times_list': [],
+    'modulations_dict': dict(),
+    'dist_block_list': [],
+    'cong_block_list': [],
+    'holding_time': None,
+    'cores_per_link': None,
+    'spectral_slots': None,
+}
